# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Restaurantes",
        "description": "Listado de endpoints de la api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crear un usuario nuevo",
                "description": "Crear un nuevo usuario en la base de datos.",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mauro Miguel"
                                    },
                                    "role_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "mauro@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "1111111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string",
                                                        "example": "success"
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Usuario registrado con exito!!!"
                                                    },
                                                    "user": {
                                                        "type": "json",
                                                        "example": "{\"name\": \"Mauro Miguel\", \"role_id\": 1, \"email\": \"mauro@email.com\", \"password\": \"11111111\"}"
                                                    },
                                                    "token": {
                                                        "type": "string",
                                                        "example": "eyJhbGciOiJIUzI1NiJ9.eyJSb2xlIjoiQWRtaW4iLCJJc3N1ZXIiOiJJc3N1ZXIiLCJVc2VybmFtZSI6IkphdmFJblVzZSIsImV4cCI6MTcwOTM0ODc1MywiaWF0IjoxNzA5MzQ4NzUzfQ.gRkdp6ZyRt60q5eQHgZXBUpC41Byy_ZnHsBUJ4dEKAw"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Iniciar sesión",
                "description": "Inicio de sesión con usuario y contraseña.",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "description": "Credenciales de inicio de sesión",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "mauro@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "11111111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string",
                                                        "example": "success"
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Login exitoso!!!"
                                                    },
                                                    "user": {
                                                        "type": "json",
                                                        "example": "{\"name\": \"Mauro Miguel\", \"role_id\": 1, \"email\": \"mauro@email.com\", \"password\": \"11111111\"}"
                                                    },
                                                    "token": {
                                                        "type": "string",
                                                        "example": "eyJhbGciOiJIUzI1NiJ9.eyJSb2xlIjoiQWRtaW4iLCJJc3N1ZXIiOiJJc3N1ZXIiLCJVc2VybmFtZSI6IkphdmFJblVzZSIsImV4cCI6MTcwOTM0ODc1MywiaWF0IjoxNzA5MzQ4NzUzfQ.gRkdp6ZyRt60q5eQHgZXBUpC41Byy_ZnHsBUJ4dEKAw"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtener el id de un usuario por su ID",
                "description": "Obtener id de un usuario",
                "operationId": "7e6eddc96aade9354b5238833ec7d950",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string",
                                                        "example": "success"
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Login exitoso!!!"
                                                    },
                                                    "user": {
                                                        "type": "json",
                                                        "example": "{\"name\": \"Mauro Miguel\", \"role_id\": 1, \"email\": \"mauro@email.com\", \"password\": \"11111111\"}"
                                                    },
                                                    "token": {
                                                        "type": "string",
                                                        "example": "eyJhbGciOiJIUzI1NiJ9.eyJSb2xlIjoiQWRtaW4iLCJJc3N1ZXIiOiJJc3N1ZXIiLCJVc2VybmFtZSI6IkphdmFJblVzZSIsImV4cCI6MTcwOTM0ODc1MywiaWF0IjoxNzA5MzQ4NzUzfQ.gRkdp6ZyRt60q5eQHgZXBUpC41Byy_ZnHsBUJ4dEKAw"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Editar usuario de un ID determinado",
                "description": "Actualizar datos usuario",
                "operationId": "703c6108dbc48db4d5682ed469585e04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Mauro Miguel"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "mauro@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "1111111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario editado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Eliminar usuario con un ID determinado",
                "description": "Eliminar usuario",
                "operationId": "cbb8757166400b4ff5f1ea2980fc6337",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
        "/api/user-pwd/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Editar contraseña de un usuario con ID determinado",
                "description": "Cambiar contraseña",
                "operationId": "9622219e2203ecea452233beda38fdc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "1111111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contruseña cambiada"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
...
    }
}